0.2.0 - Created BashpileVisitor, ThrowingErrorListener
