+++ documentation/README.md
@@ -15,6 +15,15 @@ See [Contributing](contributing/contributing.md) for helping with the project an
 * Internal Shell Layer - Makes any shell calls needed
   * May be able to change to make Linux calls directly?
 
+## Layer Diagram
+
+* Users Shell
+* CLI Layer - Handles options and arguments
+* Antlr Layer - Creates an Antlr AST
+* Engine Layer - Core logic to make the final Bash code
+* Internal Shell Layer - Makes any shell calls needed
+  * May be able to change to make Linux calls directly?
+
 ## Differences from previous version
 
 A rewrite from scratch was justified for a few reasons:
