+++ documentation/contributing/roadmap.md
@@ -1,5 +1,6 @@
-1. Setup basic program flow (ANTLR AST -> Bashpile AST -> render to Bash)
-2. `print("Hello World")`
-3. homebrew integration
-4. GHA
-5. Caching?
+1. `print("Hello World")`
+   1. Port over Bashpile.g4 files
+   2. Include m.yuvalshavit.antlr4.DenterHelper as dependency
+2. homebrew integration
+3. GHA
+4. Caching?
