+++ src/main/kotlin/org/bashpile/core/ThrowingErrorListener.kt
@@ -5,10 +5,6 @@ import org.antlr.v4.runtime.RecognitionException
 import org.antlr.v4.runtime.Recognizer
 import org.antlr.v4.runtime.misc.ParseCancellationException
 
-/**
- * Throws an exception with the same information as the default action of print to STDOUT.
- * Taken from a [Stack Overflow answer](https://stackoverflow.com/a/26573239) and auto-converted to Kotlin.
- */
 class ThrowingErrorListener: BaseErrorListener() {
     @Throws(ParseCancellationException::class)
     override fun syntaxError(
